Create multspelnet2.jl
A rewrite of multiple response glmnet package in R with sparse inputs. Underhood the package calls core functions written in C++. We use Julia instead of the R interface to deal with sparse inputs, and Julia again to port core C++ functions. The core function continues to update coefficients for each attribute, but since it is a greedy algorithm we can easily turn the sequential loop for attributes to a parallel one by first vectorizing all inner loops and memorize the minimum coefficients seen so far.